refactor: Move configurator UI components to configurator dir
