+++ app/components/chart-annotations-selector.tsx
@@ -9,7 +9,7 @@ import {
   ControlSectionContent,
   SectionTitle,
 } from "./chart-controls/section";
-import { EmptyRightPanel } from "./empty-right-panel";
+import { EmptyRightPanel } from "../configurator/components/empty-right-panel";
 import { MetaInputField } from "./field";
 import { useLocale } from "../lib/use-locale";
