+++ app/configurator/components/action-bar.tsx
@@ -5,10 +5,10 @@ import {
   canTransitionToNextStep,
   canTransitionToPreviousStep,
   useConfiguratorState,
-} from "../configurator";
-import { useDataCubeMetadataWithComponentValuesQuery } from "../graphql/query-hooks";
-import { useLocale } from "../lib/use-locale";
-import { LocalizedLink } from "./links";
+} from "..";
+import { useDataCubeMetadataWithComponentValuesQuery } from "../../graphql/query-hooks";
+import { useLocale } from "../../lib/use-locale";
+import { LocalizedLink } from "../../components/links";
 
 export const ActionBar = ({ dataSetIri }: { dataSetIri?: string }) => {
   const [state, dispatch] = useConfiguratorState();
