+++ app/configurator/components/configurator.tsx
@@ -1,13 +1,13 @@
 import { Box } from "theme-ui";
-import { useConfiguratorState } from "../../configurator";
-import { ChartAnnotationsSelector } from "../chart-annotations-selector";
-import { ChartOptionsSelector } from "../chart-options-selector";
-import { DataSetMetadata } from "../dataset-metadata";
-import { PanelLeft } from "../panel-left";
-import { PanelMiddle } from "../panel-middle";
-import { Stepper } from "../stepper";
+import { useConfiguratorState } from "..";
+import { ChartAnnotationsSelector } from "../../components/chart-annotations-selector";
+import { ChartOptionsSelector } from "../../components/chart-options-selector";
+import { DataSetMetadata } from "../../components/dataset-metadata";
+import { PanelLeft } from "./panel-left";
+import { PanelMiddle } from "./panel-middle";
+import { Stepper } from "./stepper";
 
-export const ChartEditor = () => {
+export const Configurator = () => {
   // Local state, the dataset preview doesn't need to be persistent.
   // FIXME: for a11y, "updateDataSetPreviewIri" should also move focus to "Weiter" button (?)
   const [state] = useConfiguratorState();
