+++ app/configurator/index.ts
@@ -1,3 +1,4 @@
 export * from "./config-form";
 export * from "./config-types";
 export * from "./configurator-state";
+export * from "./components/configurator";
